①seria作为基于simp协议的网络传输中的序列化组件时支持在buffer头部额外预留一个simp_header头以避免组帧时进行拷贝操作 ②seria dump操作返回的map表的键设置为std::string而非const char*，前者比较字符串本身，后者比较字符串地址，若使用后者无法进行快速的字符搜索操作 ③应用层协议simp基本框架搭建 ④server组件的连接可读时添加app_prt字段指示该连接对应的对象(bug fix) ⑤tools/registry 名字发现与注册服务完成部分开发 ⑥加入控制命令头 control_cmd 与错误统一定义头 err_define

+++ src/ini_impl.cpp
@@ -227,7 +227,7 @@ namespace crx
         auto impl = (ini_impl*)m_obj;
         auto it = impl->m_sec_map.find(sec_name);
         if (impl->m_sec_map.end() != it && impl->m_sections[it->second].valid)
-            impl->m_sec_idx = it->second;
+            impl->m_sec_idx = (int)it->second;
         else
             impl->m_sec_idx = -1;
     }
